move function to districts.ts
