+++ src/responses/districts.ts
@@ -16,6 +16,10 @@ import {
   getDayDifference,
   getStateAbbreviationByName,
 } from "../utils";
+import {
+  FrozenIncidenceData,
+  getFrozenIncidenceHistory,
+} from "../data-requests/frozen-incidence";
 
 interface DistrictData extends IDistrictData {
   stateAbbreviation: string;
@@ -370,5 +374,28 @@ export async function DistrictsRecoveredHistoryResponse(
   return {
     data,
     meta: new ResponseMeta(statesHistoryData.lastUpdate),
+  };
+}
+
+interface FrozenIncidenceHistoryData extends IResponseMeta {
+  data: {
+    [key: string]: FrozenIncidenceData;
+  };
+}
+
+export async function FrozenIncidenceHistoryResponse(): Promise<FrozenIncidenceHistoryData> {
+  const FrozenIncidenceData = await getFrozenIncidenceHistory();
+
+  const data: {
+    [key: string]: FrozenIncidenceData;
+  } = {};
+
+  for (const element of FrozenIncidenceData.data) {
+    data[element.ags] = element;
+  }
+
+  return {
+    data: data,
+    meta: new ResponseMeta(FrozenIncidenceData.lastUpdate),
   };
 }
